feat: move constants
