+++ components/Search/index.tsx
@@ -2,8 +2,8 @@ import '@algolia/autocomplete-theme-classic'
 
 import { autocomplete } from '@algolia/autocomplete-js'
 import { createQuerySuggestionsPlugin } from '@algolia/autocomplete-plugin-query-suggestions'
+import { searchClient } from 'constants/index'
 import { useEffect, useRef } from 'react'
-import { searchClient } from 'utils'
 
 const querySuggestionsPlugin = createQuerySuggestionsPlugin({
   searchClient,
