+++ constants/index.ts
@@ -1,3 +1,4 @@
+export * from './algolia'
 export * from './chains'
 export * from './contracts'
 export * from './links'
