+++ pages/api/algolia/games/index.ts
@@ -1,5 +1,5 @@
+import { algoliaIndex } from 'constants/index'
 import { NextApiRequest, NextApiResponse } from 'next'
-import { algoliaIndex } from 'utils'
 
 const index = algoliaIndex()
